move request handlers in handlers.py
